refactor() less boilerplate initialization for JsMacroFolder
