+++ src/lib.rs
@@ -1,5 +1,4 @@
-use std::collections::{HashSet};
-use serde::{Deserialize};
+use std::collections::HashSet;
 use swc_common::plugin::metadata::TransformPluginMetadataContextKind;
 use swc_core::{
     ecma::{
@@ -22,102 +21,27 @@ mod ast_utils;
 mod macro_utils;
 mod jsx_visitor;
 mod js_macro_folder;
+mod options;
 
 use builder::*;
 use ast_utils::*;
 use js_macro_folder::JsMacroFolder;
 use jsx_visitor::TransJSXVisitor;
 use crate::macro_utils::{*};
+use crate::options::{*};
 
-#[derive(Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-struct LinguiJsOptions {
-    runtime_modules: Option<RuntimeModulesConfigMap>,
-}
-
-#[derive(Deserialize, Debug, PartialEq)]
-struct RuntimeModulesConfig(
-    String,
-    #[serde(default)]
-    Option<String>,
-);
-
-#[derive(Deserialize, Debug, PartialEq)]
-#[serde(rename_all = "camelCase")]
-struct RuntimeModulesConfigMap {
-    i18n: Option<RuntimeModulesConfig>,
-    trans: Option<RuntimeModulesConfig>,
-}
-
-#[derive(Debug)]
-struct RuntimeModulesConfigMapNormalized {
-    i18n: (String, String),
-    trans: (String, String),
-}
-
-impl LinguiJsOptions {
-    fn to_options(self, env_name: &str) -> LinguiOptions {
-        LinguiOptions {
-            strip_non_essential_fields: !(matches!(env_name, "development")),
-            runtime_modules: RuntimeModulesConfigMapNormalized {
-                i18n: (
-                    self.runtime_modules.as_ref()
-                        .and_then(|o| o.i18n.as_ref())
-                        .and_then(|o| Some(o.0.clone()))
-                        .unwrap_or("@lingui/core".into()),
-                    self.runtime_modules.as_ref()
-                        .and_then(|o| o.i18n.as_ref())
-                        .and_then(|o| o.1.clone())
-                        .unwrap_or("i18n".into()),
-                ),
-                trans: (
-                    self.runtime_modules.as_ref()
-                        .and_then(|o| o.trans.as_ref())
-                        .and_then(|o| Some(o.0.clone()))
-                        .unwrap_or("@lingui/react".into()),
-                    self.runtime_modules.as_ref()
-                        .and_then(|o| o.trans.as_ref())
-                        .and_then(|o| o.1.clone())
-                        .unwrap_or("Trans".into()),
-                ),
-            },
-        }
-    }
-}
-
-#[derive(Debug)]
-struct LinguiOptions {
-    strip_non_essential_fields: bool,
-    runtime_modules: RuntimeModulesConfigMapNormalized,
-}
-
-impl Default for LinguiOptions {
-    fn default() -> LinguiOptions {
-        LinguiOptions {
-            strip_non_essential_fields: false,
-            runtime_modules: RuntimeModulesConfigMapNormalized {
-                i18n: ("@lingui/core".into(), "i18n".into()),
-                trans: ("@lingui/react".into(), "Trans".into()),
-            },
-        }
-    }
-}
 
 #[derive(Default)]
 pub struct LinguiMacroFolder {
-    options: LinguiOptions,
     has_lingui_macro_imports: bool,
-    should_add_18n_import: bool,
-    should_add_trans_import: bool,
-
     ctx: MacroCtx,
 }
 
 impl LinguiMacroFolder {
-    fn new(options: LinguiOptions) -> LinguiMacroFolder {
+    pub  fn new(options: LinguiOptions) -> LinguiMacroFolder {
         LinguiMacroFolder {
-            options,
-            ..Default::default()
+            has_lingui_macro_imports: false,
+            ctx: MacroCtx::new(options),
         }
     }
 
@@ -160,13 +84,13 @@ impl LinguiMacroFolder {
             pick_jsx_attrs(el.opening.attrs, HashSet::from(["id", "render", "comment", "context", "i18n"]))
         );
 
-        if self.options.strip_non_essential_fields {
+        if self.ctx.options.strip_non_essential_fields {
             attrs = pick_jsx_attrs(attrs, HashSet::from(["id", "render", "i18n", "context", "values", "components"]))
         }
 
-        self.should_add_trans_import = true;
+        self.ctx.should_add_trans_import = true;
 
-        let (_, trans_export) = self.options.runtime_modules.trans.clone();
+        let (_, trans_export) = self.ctx.options.runtime_modules.trans.clone();
 
         return JSXElement {
             span: el.span,
@@ -185,13 +109,13 @@ impl LinguiMacroFolder {
     }
 }
 
-impl Fold for LinguiMacroFolder {
+impl<'a> Fold for LinguiMacroFolder {
     fn fold_module_items(&mut self, mut n: Vec<ModuleItem>) -> Vec<ModuleItem> {
         let mut has_i18n_import = false;
         let mut has_trans_import = false;
 
-        let (i18n_source, i18n_export) = self.options.runtime_modules.i18n.clone();
-        let (trans_source, trans_export) = self.options.runtime_modules.trans.clone();
+        let (i18n_source, i18n_export) = self.ctx.options.runtime_modules.i18n.clone();
+        let (trans_source, trans_export) = self.ctx.options.runtime_modules.trans.clone();
 
         n.retain(|m| {
             if let ModuleItem::ModuleDecl(ModuleDecl::Import(imp)) = m {
@@ -226,11 +150,11 @@ impl Fold for LinguiMacroFolder {
 
         n = n.fold_children_with(self);
 
-        if !has_i18n_import && self.should_add_18n_import {
+        if !has_i18n_import && self.ctx.should_add_18n_import {
             n.insert(0, create_import(i18n_source.into(), quote_ident!(i18n_export[..])));
         }
 
-        if !has_trans_import && self.should_add_trans_import {
+        if !has_trans_import && self.ctx.should_add_trans_import {
             n.insert(0, create_import(trans_source.into(), quote_ident!(trans_export[..])));
         }
 
@@ -244,14 +168,7 @@ impl Fold for LinguiMacroFolder {
             return expr;
         }
 
-        let (_, i18n_export) = self.options.runtime_modules.i18n.clone();
-
-        let mut folder = JsMacroFolder {
-            strip_non_essential_fields: self.options.strip_non_essential_fields,
-            should_add_18n_import: &mut self.should_add_18n_import,
-            i18_callee_name: i18n_export.clone().into(),
-            ctx: &self.ctx
-        };
+        let mut folder = JsMacroFolder::new(&mut self.ctx);
 
         folder
             .fold_expr(expr)
@@ -265,14 +182,7 @@ impl Fold for LinguiMacroFolder {
             return expr;
         }
 
-        let (_, i18n_export) = self.options.runtime_modules.i18n.clone();
-
-        let mut folder = JsMacroFolder {
-            strip_non_essential_fields: self.options.strip_non_essential_fields,
-            should_add_18n_import: &mut self.should_add_18n_import,
-            i18_callee_name: i18n_export.clone().into(),
-            ctx: &self.ctx
-        };
+        let mut folder = JsMacroFolder::new(&mut self.ctx);
 
         folder
             .fold_call_expr(expr)
@@ -286,16 +196,9 @@ impl Fold for LinguiMacroFolder {
             return el;
         }
 
-        let (_, i18n_export) = self.options.runtime_modules.i18n.clone();
-
         // apply JS Macro transformations to jsx elements
         // before they will be extracted as message components
-        el = el.fold_with(&mut JsMacroFolder {
-            strip_non_essential_fields: self.options.strip_non_essential_fields,
-            should_add_18n_import: &mut self.should_add_18n_import,
-            i18_callee_name: i18n_export.clone().into(),
-            ctx: &self.ctx
-        });
+        el = el.fold_with(&mut JsMacroFolder::new(&mut self.ctx));
 
         if let JSXElementName::Ident(ident) = &el.opening.name {
             if self.ctx.is_lingui_ident("Trans", &ident) {
@@ -328,48 +231,4 @@ pub fn process_transform(program: Program, metadata: TransformPluginProgramMetad
     );
 
     program.fold_with(&mut LinguiMacroFolder::new(config))
-}
-
-#[cfg(test)]
-mod lib_tests {
-    use super::{*};
-
-    #[test]
-    fn test_config() {
-        let config = serde_json::from_str::<LinguiJsOptions>(
-            r#"{
-                "runtimeModules": {
-                    "i18n": ["@lingui/core", "i18n"],
-                    "trans": ["@lingui/react", "Trans"]
-                }
-               }"#
-        )
-            .expect("invalid config for lingui-plugin");
-
-        assert_eq!(config, LinguiJsOptions {
-            runtime_modules: Some(RuntimeModulesConfigMap {
-                i18n: Some(RuntimeModulesConfig("@lingui/core".into(), Some("i18n".into()))),
-                trans: Some(RuntimeModulesConfig("@lingui/react".into(), Some("Trans".into()))),
-            })
-        })
-    }
-
-    #[test]
-    fn test_config_optional() {
-        let config = serde_json::from_str::<LinguiJsOptions>(
-            r#"{
-                "runtimeModules": {
-                    "i18n": ["@lingui/core"]
-                }
-               }"#
-        )
-            .expect("invalid config for lingui-plugin");
-
-        assert_eq!(config, LinguiJsOptions {
-            runtime_modules: Some(RuntimeModulesConfigMap {
-                i18n: Some(RuntimeModulesConfig("@lingui/core".into(), None)),
-                trans: None,
-            })
-        })
-    }
 }
